Fixes eslint complains regarding old js
