+++ caseworker/advice/templates/advice/review_consolidate.html
@@ -43,4 +43,9 @@
         </div>
     </main>
 </div>
+
+{% endblock %}
+
+{% block javascript %}
+    <script src="{% static 'refusal-review-consolidate.js' %}" type="text/javascript"></script>
 {% endblock %}
