+++ caseworker/assets/styles/components/_advice.scss
@@ -220,7 +220,5 @@
 .advice-refusal-labels {
   .govuk-label {
     font-weight: 700;
-    font-family: Segoe UI, -apple-system, BlinkMacSystemFont, Arial, sans-serif;
-    -webkit-font-smoothing: antialiased;
   }
 }
