+++ package.json
@@ -5,9 +5,9 @@
   "private": true,
   "scripts": {
     "build_exporter": "parcel build ./exporter/assets/javascripts/main.js ./exporter/assets/javascripts/cookie-policy-form.js --dist-dir exporter/assets/built --public-url /assets/",
-    "build_caseworker": "parcel build ./caseworker/assets/javascripts/{main,head,bookmarks,cookie-policy-form,search-cases,search-products,beis,tau,tau-edit,case-filters}.js --dist-dir caseworker/assets/built --public-url /assets/",
+    "build_caseworker": "parcel build ./caseworker/assets/javascripts/{main,head,bookmarks,cookie-policy-form,search-cases,search-products,beis,tau,tau-edit,case-filters,refusal-review-consolidate}.js --dist-dir caseworker/assets/built --public-url /assets/",
     "watch_exporter": "parcel watch --port 8400 ./exporter/assets/javascripts/{main,head}.js ./exporter/assets/javascripts/cookie-policy-form.js --dist-dir exporter/assets/built --public-url /assets/",
-    "watch_caseworker": "parcel watch --port 8401 ./caseworker/assets/javascripts/{main,head,bookmarks,cookie-policy-form,search-cases,search-products,beis,tau,tau-edit,case-filters}.js --dist-dir caseworker/assets/built --public-url /assets/",
+    "watch_caseworker": "parcel watch --port 8401 ./caseworker/assets/javascripts/{main,head,bookmarks,cookie-policy-form,search-cases,search-products,beis,tau,tau-edit,case-filters,refusal-review-consolidate}.js --dist-dir caseworker/assets/built --public-url /assets/",
     "watch": "npm-run-all --parallel watch_exporter watch_caseworker",
     "build": "npm install --no-save && npm run build_all",
     "build_all": "export NODE_ENV=production && npm-run-all --serial build_exporter build_caseworker",
